Move all pages under pages subdirectory
