+++ premiser-ui/src/pages/propositionUsages/PropositionUsagesPage.tsx
@@ -18,8 +18,8 @@ import StatementCard from "../../StatementCard";
 import { combineIds } from "../../viewModels";
 import JustificationCard from "../../JustificationCard";
 import { useAppDispatch, useAppEntitySelector, useAppSelector } from "@/hooks";
-import ErrorPage from "@/ErrorPage";
-import AppearanceCard from "../appearances/AppearanceCard";
+import ErrorPage from "@/pages/ErrorPage";
+import AppearanceCard from "@/pages/appearances/AppearanceCard";
 import PropositionCompoundCard from "@/PropositionCompoundCard";
 import PropositionCard from "@/PropositionCard";
 import FetchMoreButton from "@/components/button/FetchMoreButton";
