+++ premiser-ui/src/pages/submitMediaExcerpt/SubmitMediaExcerptPage.tsx
@@ -1,18 +1,18 @@
+import { isArray } from "lodash";
+import queryString, { ParsedQuery } from "query-string";
 import React from "react";
 import { useDispatch, useSelector } from "react-redux";
-import queryString, { ParsedQuery } from "query-string";
 import { useLocation } from "react-router";
 
 import { CreateMediaExcerptInput } from "howdju-common";
 
 import { editors } from "@/actions";
-import { RootState } from "@/setupStore";
-import { isArray } from "lodash";
-import HowdjuHelmet from "@/Helmet";
-import ErrorMessages from "@/ErrorMessages";
-import MediaExcerptEditor from "@/editors/MediaExcerptEditor";
-import SingleColumnGrid from "@/components/layout/SingleColumnGrid";
 import { Page } from "@/components/layout/Page";
+import SingleColumnGrid from "@/components/layout/SingleColumnGrid";
+import MediaExcerptEditor from "@/editors/MediaExcerptEditor";
+import ErrorMessages from "@/ErrorMessages";
+import HowdjuHelmet from "@/Helmet";
+import { RootState } from "@/setupStore";
 
 const id = "submit-media-excerpt";
 export const editorType = "MEDIA_EXCERPT";
