+++ premiser-ui/src/routes.tsx
@@ -7,37 +7,37 @@ import { Route, RouteProps } from "react-router";
 import { commonPaths, JustificationRootTargetTypes } from "howdju-common";
 
 import AboutPage from "@/pages/about/AboutPage";
+import AccountSettingsPage from "@/pages/accountSettings/AccountSettingsPage";
+import AppearancePage from "@/pages/appearances/AppearancePage";
+import CreateAppearancePage from "@/pages/appearances/CreateAppearancePage";
+import CreatePropositionPage from "@/pages/CreatePropositionPage";
+import ExplorePage from "@/pages/explore/ExplorePage";
+import FactCheckPage from "@/pages/factChecks/FactCheckPage";
+import IconPage from "@/pages/IconPage";
+import JustificationsPage from "@/pages/justifications/JustificationsPage";
+import JustificationsSearchPage from "@/pages/justificationsSearch/JustificationsSearchPage";
+import LandingPage from "@/pages/landing/LandingPage";
+import LoginPage from "@/pages/LoginPage";
+import MainSearchPage from "@/pages/mainSearch/MainSearchPage";
+import MediaExcerptPage from "@/pages/mediaExcerpt/MediaExcerptPage";
+import MediaExcerptsSearchPage from "@/pages/mediaExcerptsSearch/MediaExcerptsSearchPage";
+import NotFoundPage from "@/pages/NotFoundPage";
 import PasswordResetConfirmationPage from "@/pages/passwordResetConfirmation/PasswordResetConfirmationPage";
 import PasswordResetRequestPage from "@/pages/passwordResetRequest/PasswordResetRequestPage";
-import CreatePropositionPage from "./CreatePropositionPage";
+import PersorgPage from "@/pages/persorg/PersorgPage";
+import PrivacySettingsPage from "@/pages/PrivacySettingsPage";
+import PropositionUsagesPage from "@/pages/propositionUsages/PropositionUsagesPage";
+import RecentActivityPage from "@/pages/recentActivity/RecentActivityPage";
+import RegistrationConfirmationPage from "@/pages/registration/RegistrationConfirmationPage";
+import RegistrationRequestPage from "@/pages/registration/RegistrationRequestPage";
+import SourcePage from "@/pages/source/SourcePage";
+import SubmitMediaExcerptPage from "@/pages/submitMediaExcerpt/SubmitMediaExcerptPage";
+import TagPage from "@/pages/tag/TagPage";
+import TestErrorPage from "@/pages/TestErrorPage";
+import WritQuotePage from "@/pages/WritQuotePage";
 import { history } from "./history";
-import IconPage from "./IconPage";
-import LandingPage from "./LandingPage";
-import LoginPage from "./LoginPage";
 import mainSearcher from "./mainSearcher";
-import NotFoundPage from "./NotFoundPage";
-import AccountSettingsPage from "./pages/accountSettings/AccountSettingsPage";
-import AppearancePage from "./pages/appearances/AppearancePage";
-import CreateAppearancePage from "./pages/appearances/CreateAppearancePage";
-import ExplorePage from "./pages/explore/ExplorePage";
-import FactCheckPage from "./pages/factChecks/FactCheckPage";
-import JustificationsPage from "./pages/justifications/JustificationsPage";
-import JustificationsSearchPage from "./pages/justificationsSearch/JustificationsSearchPage";
-import MainSearchPage from "./pages/mainSearch/MainSearchPage";
-import MediaExcerptPage from "./pages/mediaExcerpt/MediaExcerptPage";
-import MediaExcerptsSearchPage from "./pages/mediaExcerptsSearch/MediaExcerptsSearchPage";
-import PersorgPage from "./pages/persorg/PersorgPage";
-import PropositionUsagesPage from "./pages/propositionUsages/PropositionUsagesPage";
-import RegistrationConfirmationPage from "./pages/registration/RegistrationConfirmationPage";
-import RegistrationRequestPage from "./pages/registration/RegistrationRequestPage";
-import SourcePage from "./pages/source/SourcePage";
-import SubmitMediaExcerptPage from "./pages/SubmitMediaExcerptPage";
-import TagPage from "./pages/tag/TagPage";
-import WritQuotePage from "./pages/WritQuotePage";
 import paths, { createJustificationPath } from "./paths";
-import PrivacySettingsPage from "./PrivacySettingsPage";
-import RecentActivityPage from "./RecentActivityPage";
-import TestErrorPage from "./TestErrorPage";
 import { CreatePropositionPageMode } from "./types";
 
 const renderHomePath = (props: RouteProps) => {
